refactor + add comments
