+++ frontend/src/hooks/useCallbackURL.ts
@@ -1,6 +1,11 @@
 import { useQuery } from "@tanstack/react-query";
 import { getCallbackURL } from "@/api/auth";
 
+/**
+ * Provides a callback URL for authentication.
+ * 
+ * @returns The callback URL for authentication.
+ */
 export function useCallbackURL() {
   return useQuery({
     queryKey: ['callback'],
