+++ frontend/src/hooks/useCurrentUser.ts
@@ -1,7 +1,11 @@
 import { useQuery } from "@tanstack/react-query";
 import { fetchCurrentUser } from "@/api/users";
 
-
+/**
+ * Provides the current authenticated user.
+ * 
+ * @returns The current user.
+ */
 export const useCurrentUser = () => {
   return useQuery({
     queryKey: ['currentUser'],
