+++ frontend/src/pages/Classrooms/InviteTAs/index.tsx
@@ -5,7 +5,7 @@ import { useContext } from "react";
 import { SelectedClassroomContext } from "@/contexts/selectedClassroom";
 import "../styles.css";
 import { ClassroomRole } from "@/types/enums";
-import { useClassroomInviteLink } from "@/hooks/useClassroomInviteLink";
+import { useClassroomInviteLink } from "@/hooks/useClassroom";
 
 const InviteTAs: React.FC = () => {
   const { selectedClassroom } = useContext(SelectedClassroomContext);
